+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Energizing program
 - `PROGRAMS.md`
+- Cautions message
 ### Changed
 - `README.md` modified
 - `AUTHORS.md` updated